Transform tilt_x and tilt_y to stay in sync with window transforms
Tilt values from the device should be transformed before being handed to
a window in case there is any e.g. rotation or other transformation
occuring that would cause a visible disconnect between the actual and
simulated pen angle.

Note: This works pretty well, but isn't mathematically correct since
we're taking the sine of a tilt that isn't in radians (instead, its
in normalized [0..1] range).